feature-#6 [계산기] 재료 fragment를계산기 fragment에 붙이기
- fragment_calc와 fragment_calc_ingredient에 layout 태그 추가
- fragment_calc_ingredient의 margin 제거
- fragment_calc에 padding 추가
- CalcFragment에 IngredientFragment 붙임
