--- app/src/main/java/cookcook/nexters/com/amoogye/views/calc/CalcFragment.kt
@@ -1,29 +1,0 @@
-package cookcook.nexters.com.amoogye.views.calc
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import cookcook.nexters.com.amoogye.R
-
-
-class CalcFragment: Fragment() {
-
-    companion object {
-        // 선택 선언 1 (Fragment를 싱글턴으로 사용 시)
-        private var INSTANCE: CalcFragment? = null
-
-        fun getInstance(): CalcFragment {
-            if (INSTANCE == null) {
-                INSTANCE = CalcFragment()
-            }
-            return INSTANCE!!
-        }
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.fragment_calc, container, false)
-    }
-}
