+++ app/src/main/java/cookcook/nexters/com/amoogye/views/calc/presenter/CalcFragment.kt
@@ -1,0 +1,52 @@
+package cookcook.nexters.com.amoogye.views.calc.presenter
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import cookcook.nexters.com.amoogye.R
+import cookcook.nexters.com.amoogye.base.BaseFragment
+import cookcook.nexters.com.amoogye.databinding.FragmentCalcBinding
+
+
+class CalcFragment: BaseFragment() {
+    override val layoutRes: Int = R.layout.fragment_calc
+    lateinit var binding : FragmentCalcBinding
+
+    companion object {
+        // 선택 선언 1 (Fragment를 싱글턴으로 사용 시)
+        private var INSTANCE: CalcFragment? = null
+
+        fun getInstance(): CalcFragment {
+            if (INSTANCE == null) {
+                INSTANCE =
+                    CalcFragment()
+            }
+            return INSTANCE!!
+        }
+    }
+
+    override fun setupViews(view: View) {
+        val fragmentManager = fragmentManager
+        val fragmentTransaction = fragmentManager!!.beginTransaction()
+
+        val ingredientFragment = IngredientFragment()
+
+        fragmentTransaction.add(R.id.calculator_container, ingredientFragment).commit()
+    }
+
+    override fun subscribeUI() {
+    }
+
+    override fun onDataBinding(inflater: LayoutInflater, container: ViewGroup?): View? {
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_calc, container, false)
+
+        return null
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        super.onCreateView(inflater, container, savedInstanceState)
+        return inflater.inflate(R.layout.fragment_calc, container, false)
+    }
+}
